chore: index efficiency component cleanup
